refactor: move encoding to shared/ to fix reverse-direction import
Why: encoding.ts lived in document/ but was imported by shared/file-utils.ts,
creating a leak that blocks clean extraction of a core package. Encoding is
byte-level, not document-level, so shared/ is its natural home.

Prep step for the en-core/en-quire/en-scribe monorepo split (issue #58).
